Refactoring. Putting jsDocToJson into documentation.js
This is useful to keep the code related to documentation together in one place
and reducing the promiscuity with actual lib code.
