+++ test/docs-tests/generateWebsite.js
@@ -6,7 +6,7 @@ const fs = require('fs');
 const util = require('util');
 const path = require('path');
 const launchEleventy = require('./launchEleventy');
-const { jsDocToJson } = require('../../lib/jsDocToJson');
+const { jsDocToJson } = require('../../lib/documentation');
 const _ncp = require('ncp').ncp;
 _ncp.limit = 16;
 const recursiveCp = util.promisify(_ncp);
